@@ -1,63 +1,54 @@
 {
-  "name": "bergerac-roads",
-  "version": "1.0.0",
-  "description": "A great project",
-  "engines": {
-    "node": ">=10.15.0"
-  },
-  "private": true,
-  "license": "AGPL-3.0",
-  "scripts": {
-    "analyze": "nuxt build --analyze",
-    "dev": "nuxt",
-    "dev:import": "nodemon --exec babel-node data/import.js --watch data/import.js",
-    "build": "npm-run-all build:data build:app",
-    "build:data": "npm run data:import",
-    "build:app": "nuxt build --spa",
-    "start": "nuxt start --spa",
-    "data:db": "json-server data/db.json",
-    "data:import": "babel-node data/import.js",
-    "data:sample": "babel-node data/sample.js",
-    "data:sprite": "babel-node data/sprite.js",
-    "generate": "nuxt generate --spa",
-    "lint": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
-    "test": "npm-run-all --parallel dev:* test:*",
-    "test:panel": "./node_modules/.bin/cypress open",
-    "precommit": "npm run lint"
+  "author": "Romain Racamier-Lafon <romain.racamier@gmail.com>",
+  "config": {
+    "description": "Tourisme vitivinicole dans le Bergerac",
+    "title": "Vins de Bergerac Duras"
   },
   "dependencies": {
-    "@nuxtjs/pwa": "2.6.0",
+    "@nuxtjs/pwa": "3.0.0-beta.14",
     "axios": "0.18.0",
-    "babel-cli": "6.26.0",
-    "babel-eslint": "8.2.6",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-env": "1.7.0",
+    "core-js": "2",
     "element-ui": "2.7.2",
     "fast-deep-equal": "2.0.1",
     "json-server": "0.12.2",
     "lodash": "4.17.11",
     "node-sass": "4.11.0",
-    "npm-run-all": "4.1.5",
-    "nuxt": "1.4.5",
-    "nuxt-sass-resources-loader": "2.0.5",
+    "nuxt": "2.6.1",
     "sass-loader": "7.1.0",
     "speakingurl": "14.0.1",
-    "style-loader": "0.21.0",
     "vue-i18n": "7.8.1",
     "vue-lazyload": "1.2.6",
-    "vue-multiselect": "2.1.4",
     "vue2-google-maps": "0.9.8"
   },
+  "description": "A great project",
   "devDependencies": {
-    "@types/node-sass": "3.10.32",
-    "eslint": "4.19.1",
-    "eslint-friendly-formatter": "4.0.1",
+    "@nuxtjs/eslint-config": "0.0.1",
+    "babel-eslint": "10.0.1",
+    "eslint": "5.15.1",
+    "eslint-config-standard": "12.0.0",
     "eslint-loader": "2.1.2",
-    "eslint-plugin-vue": "4.7.1",
-    "glob-all": "3.1.0",
-    "is-installed": "2.0.1",
-    "node-sprite-generator": "0.10.2",
-    "nodemon": "1.18.11",
-    "stylelint-config-standard": "18.2.0"
-  }
+    "eslint-plugin-import": "2.16.0",
+    "eslint-plugin-jest": "22.3.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-nuxt": "0.4.2",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
+    "eslint-plugin-vue": "5.2.2"
+  },
+  "engines": {
+    "node": ">=10.15.0"
+  },
+  "license": "AGPL-3.0",
+  "name": "bergerac-roads",
+  "private": true,
+  "scripts": {
+    "analyze": "nuxt build --analyze",
+    "build": "npm run data:import && nuxt build",
+    "data:db": "json-server data/db.json",
+    "data:import": "node data/import.js",
+    "dev": "nuxt",
+    "lint": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
+    "precommit": "npm run lint"
+  },
+  "version": "2.0.0"
 }
