@@ -1,88 +1,86 @@
-module.exports = {
+import pkg from './package'
+
+export default {
+  mode: 'spa',
+
   /*
   ** Headers of the page
   */
   head: {
-    title: 'Vins de Bergerac Duras',
+    title: pkg.config.title,
     htmlAttrs: {
       lang: 'fr-FR',
     },
     meta: [
       { charset: 'utf-8' },
       { name: 'viewport', content: 'width=device-width, initial-scale=1' },
+      { hid: 'description', name: 'description', content: pkg.description },
       { name: 'theme-color', content: '#ffffff' },
-      {
-        hid: 'description',
-        name: 'description',
-        content: 'Tourisme vitivinicole dans le Bergerac',
-      },
     ],
     script: [
       { src: '/scripts/rollbar.js', defer: true },
     ],
+    link: [
+      { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' },
+    ],
     noscript: [{ innerHTML: 'Ce site nécessite JavaScript.' }],
   },
-  css: [
-    '~assets/styles/shared/global.scss',
-    'element-ui/lib/theme-chalk/index.css',
-    'element-ui/lib/theme-chalk/display.css',
-    'vue-multiselect/dist/vue-multiselect.min.css',
-  ],
   env: {
     api: 'http://localhost:3003',
     cdn: 'https://atg502nfn.cloudimg.io/cdn/none/none/bergerac.lebowsky-dev.xyz',
-    cdnBase: 'https://atg502nfn.cloudimg.io/cdn/none/none/'
+    cdnBase: 'https://atg502nfn.cloudimg.io/cdn/none/none/',
   },
   /*
   ** Customize the progress bar color
   */
   loading: { color: '#3B8070' },
   /*
+  ** Global CSS
+  */
+  css: [
+    '@/assets/styles/shared/global.scss',
+    'element-ui/lib/theme-chalk/display.css',
+    'element-ui/lib/theme-chalk/index.css',
+  ],
+
+  /*
+  ** Plugins to load before mounting the App
+  */
+  plugins: [
+    '@/plugins/element-ui',
+    '@/plugins/db',
+    '@/plugins/i18n',
+    '@/plugins/lazyload',
+    '@/plugins/google-maps',
+  ],
+
+  /*
+  ** Nuxt.js modules
+  */
+  modules: [
+    '@nuxtjs/pwa',
+  ],
+
+  /*
   ** Build configuration
   */
   router: {
     middleware: 'i18n',
   },
-  plugins: [
-    'plugins/element-ui',
-    'plugins/db',
-    'plugins/i18n',
-    'plugins/lazyload',
-    'plugins/google-maps',
-  ],
-  modules: [
-    [
-      'nuxt-sass-resources-loader',
-      {
-        resources: ['./assets/styles/ressources/variables.scss', './assets/styles/ressources/icons.scss', './assets/styles/ressources/icons-transparent.scss'],
-      },
-    ],
-    '@nuxtjs/pwa',
-  ],
   build: {
-    vendor: ['babel-polyfill', 'element-ui'],
+    transpile: [/^element-ui/],
+
     /*
-    ** Run ESLint on save
+    ** You can extend webpack config here
     */
     extend(config, ctx) {
-      // Exécuter ESLint lors de la sauvegarde
+      // Run ESLint on save
       if (ctx.isDev && ctx.isClient) {
         config.module.rules.push({
           enforce: 'pre',
           test: /\.(js|vue)$/,
           loader: 'eslint-loader',
           exclude: /(node_modules)/,
-        })
-      }
-      if (!ctx.isClient) {
-        // This instructs Webpack to include `vue2-google-maps`'s Vue files
-        // for server-side rendering
-        config.externals.splice(0, 0, function(context, request, callback) {
-          if (/^vue2-google-maps($|\/)/.test(request)) {
-            callback(null, false)
-          } else {
-            callback()
-          }
         })
       }
     },
