@@ -1,11 +1,13 @@
 <template>
   <el-container direction="vertical" class="admin-table">
     <h2>Domaines et chateaux</h2>
-    <el-table :data="domains" stripe :style="'width: 100%'" v-loading="loading" :row-class-name="tableRowClassName">
-      <el-table-column type="selection" width="55"/>
-      <el-table-column prop="title" label="Nom"/>
+    <el-table :data="domains" stripe style="width: 100%" v-loading="loading" :row-class-name="tableRowClassName">
+      <el-table-column type="selection" width="55" />
+      <el-table-column prop="title" label="Nom" />
       <el-table-column label="Actif ?" width="180">
-        <template slot-scope="scope">{{ scope.row.active ? 'oui' : 'non' }}</template>
+        <template slot-scope="scope">
+          {{ scope.row.active ? 'oui' : 'non' }}
+        </template>
       </el-table-column>
     </el-table>
   </el-container>
@@ -28,7 +30,7 @@ export default {
       this.loading = true
       fetch(this.api)
         .then(response => response.json())
-        .then(domains => {
+        .then((domains) => {
           this.domains = domains
           this.loading = false
         })
@@ -56,6 +58,8 @@ export default {
 }
 </script>
 <style lang="scss" scoped>
+@import "@/assets/styles/ressources/variables.scss";
+
 .admin-table .el-table {
   background-color: $white;
 }
