@@ -1,7 +1,7 @@
 <template>
   <div class="footer">
     <small>{{ $t('footer.alcoolWarning') }}</small>
-    <div class="grappe"/>
+    <div class="grappe" />
     <small>{{ $t('footer.copyright') }}</small>
     <small class="line">
       <nuxt-link :to="$i18n.path('legal')">{{ $t('legal.title') }}</nuxt-link>
@@ -12,6 +12,8 @@
 </template>
 
 <style lang="scss">
+@import "@/assets/styles/ressources/variables.scss";
+
 .footer {
   padding: 30px 0 20px;
   background-color: $rouge;
