@@ -1,7 +1,6 @@
 <template>
   <el-dropdown class="vineyards" v-loading="loading">
-
-    <span class="el-dropdown-link">{{ $t('vineyards.titles.ours') }}<i class="el-icon-arrow-down el-icon--right"/></span>
+    <span class="el-dropdown-link">{{ $t('vineyards.titles.ours') }}<i class="el-icon-arrow-down el-icon--right" /></span>
     <el-dropdown-menu slot="dropdown">
       <nuxt-link v-for="vineyard in vineyards" :key="vineyard.id" :to="$i18n.path('vignoble/' + vineyard.name)">
         <el-dropdown-item>{{ $t('vineyards.' + vineyard.name + '.title') }}</el-dropdown-item>
@@ -25,7 +24,7 @@ export default {
     init() {
       this.loading = true
       console.log('Header Vineyards : init')
-      this.$db.getVineyards().then(vineyards => {
+      this.$db.getVineyards().then((vineyards) => {
         console.log('Header Vineyards : got ', vineyards)
         this.vineyards = vineyards
         this.loading = false
@@ -36,6 +35,8 @@ export default {
 </script>
 
 <style lang="scss" scoped>
+@import "@/assets/styles/ressources/variables.scss";
+
 .el-dropdown {
   height: 20px;
   color: $white;
