@@ -1,5 +1,5 @@
 <template>
   <div>
-    <nuxt/>
+    <nuxt />
   </div>
 </template>
