@@ -1,6 +1,7 @@
 import Vue from 'vue'
 import Element from 'element-ui'
+import locale from 'element-ui/lib/locale/lang/en'
 
-const locale = require('element-ui/lib/locale/lang/fr')
-
-Vue.use(Element, { locale })
+export default () => {
+  Vue.use(Element, { locale })
+}
