@@ -7,7 +7,7 @@ function getDomainUrl(domain) {
   return 'domaine/' + domain.id + '-' + getSlug(domain.title)
 }
 
-eventBus.$on('goto-domain', domain => {
+eventBus.$on('goto-domain', (domain) => {
   const url = getDomainUrl(domain)
   $nuxt.$router.push(url)
 })
